pyproject.toml
[project]
name = "jax-examples"
version = "0.1.0"
description = "Examples using the JAX framework"
requires-python = ">=3.10"
dependencies = [
    "arviz>=0.22.0",
    "blackjax>=1.2.5",
    "cvxpy>=1.7.2",
    "cvxpylayers>=0.1.9",
    "diffrax>=0.7.0",
    "equinox>=0.13.0",
    "flax>=0.10.7",
    "ipython>=8.37.0",
    "jax>=0.6.2",
    "jupyter>=1.1.1",
    "keras>=3.11.3",
    "matplotlib>=3.10.6",
    "optax>=0.2.5",
    "pandas-stubs>=2.3.2.250827",
    "protobuf==3.20.3",
    "pymc>=5.25.1",
    "python-dotenv>=1.1.1",
    "rdatasets>=0.2.10",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "tensorflow>=2.19.1",
    "tensorflow-datasets>=4.9.9",
    "tensorflow-probability>=0.25.0",
    "ydf>=0.9.0",
]
